tests: all TDD tests ok

+++ src/main/java/cat/itacademy/s04/t02/n02/services/ProviderServiceImpl.java
@@ -1,11 +1,13 @@
 package cat.itacademy.s04.t02.n02.services;
 
 import cat.itacademy.s04.t02.n02.exceptions.DuplicateProviderName;
+import cat.itacademy.s04.t02.n02.exceptions.ProviderHasFruitsException;
 import cat.itacademy.s04.t02.n02.exceptions.ProviderNotFoundException;
 import cat.itacademy.s04.t02.n02.model.Fruit;
 import cat.itacademy.s04.t02.n02.model.FruitDTO;
 import cat.itacademy.s04.t02.n02.model.Provider;
 import cat.itacademy.s04.t02.n02.model.ProviderDTO;
+import cat.itacademy.s04.t02.n02.repository.FruitRepository;
 import cat.itacademy.s04.t02.n02.repository.ProviderRepository;
 import org.springframework.stereotype.Service;
 
@@ -13,9 +15,11 @@ import java.util.List;
 
 @Service
 public class ProviderServiceImpl implements ProviderService {
+	private final FruitRepository fruitRepository;
 	private final ProviderRepository providerRepository;
-	public ProviderServiceImpl(ProviderRepository providerRepository){
+	public ProviderServiceImpl(ProviderRepository providerRepository, FruitRepository fruitRepository){
 		this.providerRepository = providerRepository;
+		this.fruitRepository = fruitRepository;
 	}
 	public ProviderDTO providerToProviderDTO(Provider provider){
 		return new ProviderDTO(
@@ -58,6 +62,9 @@ public class ProviderServiceImpl implements ProviderService {
 	public void deleteProviderById(long id){
 		Provider foundProvider = providerRepository.findById(id)
 				.orElseThrow(()-> new ProviderNotFoundException("Provider with id " + id + " not found"));
+		if (fruitRepository.fruitsExistsByProviderId(id)) {
+			throw new ProviderHasFruitsException("Cannot delete provider with id " + id + " because it has associated fruits.");
+		}
 		providerRepository.delete(foundProvider);
 	}
 }
